Deleted extra gen-nix folder along with uncomment code in parser.py for main branch
