+++ curl_ex/flake.nix
@@ -1,0 +1,23 @@
+{
+  inputs =  {
+    nixpkgs.url = "github:NixOS/nixpkgs/8cad3dbe48029cb9def5cdb2409a6c80d3acfe2e";
+  };
+  description = "Basic template flake" ;
+  outputs = { self, nixpkgs }:
+    let
+      system = "x86_64-linux";
+      pkgs = import nixpkgs { inherit system; config.allowUnfree = true; };
+    in
+      {
+        # Here we are creating a basic shell w/ the program installed as a binary
+        devShells."${system}".default =
+          let
+            staticTools = with pkgs; [
+              # Add static analysis tools here:
+              pkgs.cppcheck
+            ];
+          in
+          pkgs.curl.overrideAttrs (prev: { nativeBuildInputs = (prev.nativeBuildInputs or []) ++ staticTools;});
+      };
+}
+
