--- cve-parsing/test.py
@@ -1,2 +1,0 @@
-import tensorflow as tf
-print(tf.__version__)
