--- gen-nix/scrape-site.py
@@ -1,29 +1,0 @@
-# TODO given a range of versions - find a version that would work from:
-# https://lazamar.co.uk/nix-versions/?channel=nixpkgs-unstable&package=<insert-pkg-here>
-# to see which revision would work
-
-import requests
-from bs4 import BeautifulSoup
-
-def get_packagelist(url):
-    page = requests.get(url)
-
-    soup = BeautifulSoup(page.content, "html.parser")
-    packageVersionList = []
-    for tr in soup.findAll('table')[0].findAll('tr'):
-        if tr.find('td'):
-            package = tr.findAll('td')[0].contents[0]
-            version = tr.findAll('td')[1].contents[0]
-            revision = tr.findAll('td')[2].contents[0].string
-            packageVersionList.append([package, version, revision])
-        
-    versionToRevisionMap = {}
-    for row in packageVersionList:
-        versionToRevisionMap[row[1]] = row[2]
-
-    return versionToRevisionMap
-
-if __name__ == "__main__":
-    URL = "https://lazamar.co.uk/nix-versions/?channel=nixpkgs-unstable&package=curl"
-    list = get_packagelist(URL)
-    print(list)
