--- gen-nix/service_setting_from_json.py
@@ -1,41 +1,0 @@
-from pprint import pprint
-import json
-import argparse
-
-
-def format_as_str(data):
-    settings = ""
-    for key, val in data.items():
-        # skip non-string vals
-        if key in ["package", "dataDir"]:
-            continue
-        if val is None:
-            val = "null"
-        elif isinstance(val, bool):
-            val = "true" if val else "false"
-        elif isinstance(val, str):
-            val = f'"{val}"'
-        new_setting = "{} = {};".format(key, val)
-        settings += new_setting + "\n"
-    return settings
-
-# read the local options.json file
-# get all entries contain the key
-def setting_from_json(serveice_key, file_name="options.json"):
-    with open(file_name, 'r') as file:
-        data = json.load(file)
-    relevant_data = {}
-    for key, val in data.items():
-        if serveice_key in key:
-            clean_key = key.replace(serveice_key + ".", "")
-            relevant_data[clean_key] = val['default']
-    print(relevant_data)
-    return format_as_str(relevant_data)
-
-# accessKey = "";
-# browser = true;
-#
-
-if __name__ == '__main__':
-    result = setting_from_json("services.minio")
-    print(result)
