--- gen-nix/templates/basic.nix
@@ -1,20 +1,0 @@
-{
-  inputs =  {
-    nixpkgs.url = $url;
-  };
-  description = "Basic template flake" ;
-  outputs = { self, nixpkgs }:
-    let
-      system = "x86_64-linux";
-      pkgs = import nixpkgs { inherit system; };
-    in
-      {
-        # Here we are creating a basic shell w/ the program installed as a binary
-        devShells."$${system}".default = pkgs.mkShell {
-          buildInputs =  with pkgs; [
-            $pkg
-          ];
-        };
-      };
-}
-
