--- gen-nix/templates/build-and-tool-env.nix
@@ -1,23 +1,0 @@
-{
-  inputs =  {
-    nixpkgs.url = $url;
-  };
-  description = "Basic template flake" ;
-  outputs = { self, nixpkgs }:
-    let
-      system = "x86_64-linux";
-      pkgs = import nixpkgs { inherit system; config.allowUnfree = true; };
-    in
-      {
-        # Here we are creating a basic shell w/ the program installed as a binary
-        devShells."$${system}".default =
-          let
-            staticTools = with pkgs; [
-              # Add static analysis tools here:
-              $statictools
-            ];
-          in
-          pkgs.$pkg.overrideAttrs (prev: { nativeBuildInputs = (prev.nativeBuildInputs or []) ++ staticTools;});
-      };
-}
-
